changes: + added section <architectures> <arch .../> </architectures> to specify the list of architectures for the installation source. This is used instead of the arch="..." attribute of the <instsource> container which is now forced empty. The architectures list must be set exactly once and contain at least one <arch> element. The fallback attribute is optional, omitting means "no fallback" implicitely.
+ added section <chroot package=.../> in metadata section. It may be
required to set up a chroot environment within the instsource creation
process to run shellscripts on metapackages and metafiles. For the
latter there is yet no mechanism to specify its requirements -- now
here it comes.

+ For some steps of the metadata creation it is necessary to provide
additional information. In the current Autobuild world this is done by
shell variables. This is implemented here using the <productinfo
name=...">...</productinfo> container within the <metadata> container.
It is possible to set name=value pairs just like setting a shell
variable.
These changes should not violate the schemeversion 2.0 since noone uses
 the <instsource> yet.

+ the handling method "getInstSourceArchList" works differently now due
to changed structure and more information; also a new method
"getInstSourceProductInfo" is available now. Implementation of the
<chroot> will follow soon


git-svn-id: https://svn.berlios.de/svnroot/repos/kiwi/kiwi-head@1135 92bc7f66-5d17-0410-ac0d-e081219a1bb3