Fix the handling of the --mbrid command line switch
The value passed to the --mbrid flag is parsed to an integer by
GetOptions (format is 'o' = extended integer). There is code checking
whether this integer is within the expected range and formats it as
hexadecimal value.

But a5c7ab512dfb02dd99044c1db1cbc46ab05bf221 introduced a regression
in the way the value of the --mbrid flag was handled because the value
is set using $cmdL -> setMBRID before the range checking and conversion.

Signed-off-by: Justus Winter <winter@pre-sense.de>