Allow preconfiguring of s390 network cards
The default configuration DASD 191 is checked for a parm file with the
name <LINUX_USER_ID>.PARM-S11.
If it is found the parameters are exported and used to set up the
network with setupNetworkInterfaceS390.
If the file "/etc/deactivate_s390_network_config_from_dasd" is found the
check for the network configuration isn't executed. This file is removed
afterwards.