Make kiwi work again in Factory
various commands have been moved to /usr/sbin, which
is not in path for kiwi, so KIWILocator does not find it.

Also "which" got splitted into a separate package, which
is not required by kiwi, so it might not be installed.
Simply use "type -p", which is a bash builtin and does the
job as well.