Don't use mkinitrd for EC2 simple storage images
In the past we used mkinitrd to create the initrd to boot an
ec2 instance. But mkinitrd has evolved into a direction which
does not allow to create it offline, meaning not running it on
the target system itself. Same applies to dracut. Thus this
commit changes the code to let kiwi natively build its own
initrd using the vmxboot boot code to boot ec2 instances