drop .channel files for products, generate .report files instead
The approach to generate _channel file for OBS maintenance is not working,
since it requires the knowledge of OBS setup internas. There we drop this
here, let kiwi generate instead report files which document which files got
used.

We can use them for diffing results, for tracking build results in OBS and
obs worker can generate _channel files with knowledge of setup internas.