Added support for creating vagrant boxes
In order to create a vagrant box the format parameter in the
image type acts as trigger to start the box creation. Vagrant
boxes requires some metadata in order to allow adding the box
for a specific provider. Thus the following minimal type spec
is needed e.g:

  <type image="vmx" ... format="vagrant">
    <vagrantconfig provider="libvirt" virtualsize="16"/>
  </type>

The result .box file can be added to vagrant as follows

  vagrant box add <name> /path/to/file.box

Once added vagrant can run the box using the specified provider
Please note the kiwi image for vagrant needs to provide some
configuration like the existence of a user vagrant and ssh
keys. All this can be achieved in a kiwi image description.
For details visit:

  http://docs.vagrantup.com/v2/boxes/base.html