Allow to boot LVM images without extra boot partition
Added new type attribute bootpartition=true|false. The value
controls the creation of a boot partition for disk images and
overwrites the default layout kiwi would create if this is not
specified. It's not guaranteed that the resulting image is able
to boot with regards to the underlaying bootloader capabilities
Given that grub2 in a recent version is used booting LVM images
using extX and btrfs filesystems will work