Add a post install script for kiwi
The file /dev/shm/lwp-download will be removed in order to
allow kiwi to create this file with the permissions it needs.
The support to allow kiwi to run kiwi --info as non root user
requires this cleanup step because lwp-download is created
with permissions which allows root and normal users to call
the script. former versions leaves a version of the script
which doesn't allow non root users to call or change it which
results in a raise condition for a non privileged --info
call