Eliminate use of haveBootPartition variable in the initrd code
kiwi provides exact information in the partition ID's variables
created when the image is build in setupPartIDs. The indicator if
a boot partition is in use was not set correctly by the existing
function. The indicator for a boot partition should be the
filesystem label set to 'BOOT' and the check if this partition is
not equal to the root partition. In the repartition code the
partition ID for the root partition is always defined correctly
by the kiwi_RootPart variable and should not depend on any
condition. That way there was no code left which requires and
extra boolean haveBootPartition variable and also fixes the wrong
pID assignment described in bug (bnc #897497)