Fixed double encoded luks system (bnc #898249)
when kiwi builds a disk image it can either dump a filesystem
image onto a raw partition or it rsync a tree of files onto
a partition which has a filesystem created on top of a subsystem
or a subsystem stack like lvm, luks, etc. In kiwi this is
controlled by the value of the $treeAccess variable. If
treeAccess is disabled but the support for syncing a tree in
the given configuration is implemented this results in
duplicate preparation of the filesystem and the subsystems
which results in an unbootable mess.