Added support for custom boxname for vagrant images (bnc #896205)
By default the name set to the vagrant .json file is the name
of the image as specified in the kiwi XML description. However,
in the world of https://vagrantcloud.com/, the convention for
"fully-qualified" box names is "ORGANIZATION/BOXNAME" which does
not apply to a standard kiwi image name. With the new boxname
attribute people can now specify the name of the box to be
vagrantcloud conform. e.g

   <vagrantconfig ... boxname="opensuse/minimal-13.1"/>