Use btrfs as write partition for hybrid images (bnc #900922)
For hybrid live iso images there is the optional support to
create a write partition for data persistency. The initial
creation of the filesystem can take a long time on e.g USB
sticks. The creation of a btrfs filesystem without static
inode tables is much faster. Thus the switch to btrfs