Fixed and cleanup unit test processing
When running the unit tests the class KIWILog is overwritten
in a way that a global kiwi logger is created from Common::ktLog
but there is no declaration which really overwrites the KIWILog
package to be a Common::ktLog. This patch fixes this and makes
unit testing work in travis again. In addition methods from
Common::ktLog no longer present in KIWILog were removed too