Fixed suseStripKernel for updates/ modules
If the dependency checker finds a dependency to a module in the updates
or weak-updates directory the move of the module will fail because kiwi
always uses updates/weak-updates modules and has moved them already.
The dependency code however creates a new updates/weak-updates directory
for the affected module and that causes a conflict when the original
updates/weak-updates are moved back to the final place. In the end the
result is that no updates/weak-updates module are present in the initrd