Added fribidi and fbiterm to kiwi-media-requires package
Unfortunately these two packages are not on the openSUSE 13.2
DVD. But kiwi uses fbiterm in the boot images. Thus users
have to add the online repo in order to build. For 13.2 it's
too late but with this change I hope the two packages will
be placed on the DVD again