Set default bootloader to grub2
A new schema version 6.2 has been introduced in order to
check for the bootloader attribute. If not set the bootloader
is set to grub2. suboptimal bootloader settings based on
arch and/or firmware in the code has been removed. The
consistency of the bootloader setup will be handled by
runtime checks exclusively