Fixed setupBTRFSSubVolumes
if a subvolume for e.g /var/spool should be created kiwi
failed because the /var path did not exist. As a workaround
people could add /var as a volume but if only /var/spool
should be a volume and not /var there was no way to do it.
This commit fixed it in a way that kiwi checks if the path
to the volume exists and if not it is created prior to the
creation of the subvolume