Adapt changes in overlayfs in kernel 3.18
Starting with overlayfs v22 the workdir mount option is required.
In addition the module has been renamed from overlayfs to overlay
which also impacts the mount options. In order to stay compatible
kiwi implements a two way mount call, first with the legacy mount
options and second with the current mount options