Eliminate grub2-efi path
At the beginning of grub2 the home in boot was set to grub2-efi
as well as some tools take this information in their name. It's
time to cleanup this. In addition kiwi now creates a grub.cfg
which is able to check for grub_platform in order to distinguish
between efi and non efi configuration differences if there are
any