Use udevadm control to exit udev daemon
Killing the udevd is done in two steps. First we issue
an udevadm call to exit udevd the documented way. According
to the man page the call waits until udevd is stopped.
After the call we check against the process ID of udevd
we know about if they really have been stopped. If not we
kill them