Handle stop of udev daemon correctly
Killing the udevd is done in two steps. First we issue
an udevadm call to exit udevd the documented way. According
to the man page the call waits until udevd is stopped.
After the call we check against the process ID(s) of udev
daemons we know about and if they are not stopped we kill
them.