Fixed creation of split images
The filename for the splitfile which is the read-write portion
of the image was build taking the displayname into account.
Only if the displayname was not set the filename was correct
because in that case the displayname and the diskname are the
same. If a custom displayname is set the splitfile name was
wrong and the read-write data was not included to the image

In addition this patch uses the kiwi.buildinfo file to check
for the image type instead of taking weird assumptions based
on file names