Fixed callPartitioner
Instead of making the decision for the correct partition tool
by the name of the device use a method which knows better about
the type of the storage device. This applies to s390 only were
a dasd device could be a real dasd with 4k blocksize or an
emulated dasd with 512byte blocksize. In the first case fdasd
is used and in the other case parted must be used. The decision
for the tool is now based on the result if isDASDDevice()
which returns false for emulated dasd devices because we
only distinguish a different handling if the blocksize of the
device is _not_ 512byte