Added support for s390 emulated DASD devices
An emulated DASD device is a disk with 512byte blocksize configured
with the standard dasd_configure routine. In order to allow zipl to
boot from the device kiwi assume the disk mode to be 'LDL' for 4k
disks and 'SCSI' for 512b disks. In case of emulated dasd devices
the mode must be 'FBA' and there are also CDL 4k DASD disks. This
means the assumption of the later target could be wrong which is
the reason why this patch also add the

   zipl_targettype="CDL | LDL | FBA | SCSI"

attribute to the kiwi XML configuration. This fix is related to the
bugzilla reports (bnc #912199), (bnc #902424) and (bnc #898249)