Fixed zero block analyser
If all free blocks in a filesystem contain a zero the compression
result of the later image is better compared to if there is any
random junk in it. In order to have all free blocks set to a zero
byte kiwi used the dd + rm method which had the disadvantage of
destroying the sparseness of the image file. Because of that the
method was applied only if the image was built from a real disk
device which did not allow us to strip away a sparse image from
anyways. However this is not very flexible because:

1. In most cases an image is built from a loop device mapped
   against a sparse file. Thus the method is not called most
   of the time

2. With filesystems using more complex internal structures like
   xfs btrfs, zfs and friends there is not much difference in
   the compression results, but the time to create a huge empty
   file which is removed later is signifcantly recognized

Thus this change deletes the old way of zero out free blocks and
uses a tool to do the job for us. For extX based filesystems the
zerofree utility exists and is used if installed. For other
filesystems I'm not aware of a tool and thus we don't handle
them at the moment. If there is tool available kiwi print a
warning message