Fixed restoreBtrfsSubVolumes
if a subvolume for e.g /var/crash should be restored, kiwi
failed because the /var path did not exist. As a workaround
people could add /var as a volume but if only /var/crash
should be a volume and not /var there was no way to do it.
This commit fixed it in a way that kiwi checks if the path
to the volume exists and if not it is created prior to the
restore of the subvolume