Changed validation in OEM partition install mode
In partition based install kiwi checks if the current partitions
can be used for installation. Part of the validation was to check
if a filesystem signature could be found. If yes the partition
was not offered for installation. However this test is too strict
because in order to remove the filesystem signature the filesystem
must be actively destroyed. It's better to check for the filesystem
UUID instead. In order to allow the installer to use the partition
one can clear the uuid with the appropriate filesystem utility
e.g

  e2fstune -U clear

for extX based filesystems and that will not destroy the filesystem
integrity. This means an active system can be prepared for install
by clearing the uuid instead of self destructing it