Fixed setup of GPT partition table
kiwi uses parted to create partition tables. In case of GPT
a linux default partition table has a guid set which points
to a Microsoft base partition. This is wrong and I did not
find a way to correct this with parted. Therefore the sgdisk
utility will be used to correct the partition code so that
we can at least see the right values when using sgdisk