Fixed partedGetPartitionID
In case of a GPT parted has no information about a partition id
Therefore we mapped the id to the known msdos values according
to the name kiwi defines for the partition. Any GPT specific
partition got the same id 'ee' which results in not being able
to distinguish them. However for the partition based install
in EFI mode it's required to identify any potential target install
partition by an id. Therefore partedGetPartitionID will return
the sgdisk partition code and will only map to the old legacy
values if sgdisk is not available