Ensure recovery entry is appended for elilo compat
elilo does not support adding custom entries to the written
grub configuration. Thus we have to make sure the recovery
entry doesn't get lost after a recovery run. Therefore the
menu entry is added as part of the installBootloader function
because this function is called after a recovery run while
the installBootloaderRecovery functions are only called once
at initialization time of the recovery partition