Fixed use of ovftool v3.5.x
If format="ova" is specified kiwi uses VMware's ovftool if installed
on the host to generated the ova tarball. In addition kiwi extracts
the vmdk and ovf metadata from the ovftool generated data and overwrites
it with the files created by kiwi. This allows us to create vmdk
versions like VSphere needs it