Fixed console initialization on exception
If multiple consoles are used like console=tty0 console=ttyS0
the output is multiplexed among all consoles if plymouth or blogd
is in use. However if a system shell exception occurs the
exception shell was not correctly initialized as controlling
terminal. In order to fix this the sulogin program is used and
the shell .bashrc will reset the spawned console first.
This fixes (bnc#928090)