Added target_blocksize attribute in type section
In preparation to support custom blocksize values for loops
the attribute target_blocksize was implemented like the folowing
example shows:

<type ... target_blocksize="4096"/>

This will set a 4k blocksize used for all loops setup during
image build and on first boot of the image. The actual code to
apply the blocksize to the loop is not yet implemented because
we are still waiting for the loop driver to support it. This is
related to (bnc#931634)