Fixed grub2 setup on btrfs without boot partition
If the image is built with btrfs and no extra boot partition is requested
the grub2 btrfs module is used to read the boot data from the filesystem.
However in kiwi we setup a btrfs filesystem in a way that all subvolumes
are nested below a master '@' volume. Within the scope of grub2 it's
required to adapt any path in the filesystem to match this because it's
not possible to specify a subvolume set-default here. Because of this
/boot/grub2/grub.cfg becomes /@/boot/grub2/grub.cfg within grub2.
This fixes (bnc#933664)