Revert "Implement setup of loop blocksize"
The user space interface will change because in order to set the
pyhsical blocksize the value must be present on loop creation. This
required options to the losetup call and can't be done afterwards.
Therefore the implementation in kiwi will be reverted until the
user space interface for losetup is final

This reverts commit d2dfead392a7dd2064afbd59df4c4d67bcacbc98.