Fixed legacy grub install for install stick images
Installing legacy grub can happen using the image grub or the
build host grub. However for install stick disks there is no
grub in the install image even if it is present in the system
image the stick installs. Therefore the host installed grub
should be used as a fallback. This Fixes #471