Added __checkImageIncludeApplicable runtime check
The imageinclude attribute is currently not supported with all
URI types. Therefore we test if the repo URL uses obs|iso and
provide a message to the user. If we want to support this in the
future for obs URI types the imageinclude feature has to auto
map the obs URI into the correct http URI which will only work
for repositories in the public part of the open buildservice
This Fixes #472