Deleted yum grouplist check
Prior to installing yum groups, kiwi performed a grouplist
to check if the group exists. That's because yum groupinstall
does only print a warning message if a non existing group
should be installed. The problem here is that the group names
displayed by yum grouplist are the description names like
"Console internet tools" but the group id name in this case
"console-internet" is not displayed. This makes it impossible
for kiwi to use the yum group id names which we should prefer
over the description texts, because the texts could contain
special characters which makes it hard to validate the content
in a stable way. In addition yum groupinstall works more
reliable if the group id names are used