Prevent command variables for Xen domU grub2 setup
Xen PV images which boot via pvgrub have the problem that pvgrub
is not able to correctly read in the kiwi written grub.cfg file.
This is because kiwi uses a variable which contains the loader
command e.g $linux instead of the loader command directly. grub2
supports this but pvgrub is not able to interpret this information.
Therefore this patch prevents the use of the variable if the
target image is Xen, domU and the firmware type is set to bios