Prevent prefix setup in grub.cfg for Xen
This is because Xen PV guests boot via a first stage loader pygrub/pvgrub
and interpret the grub config file differently. One inconsistency is that
pvgrub searches for the grub modules at a different place. Setting up the
prefix will point pvgrub to the wrong place and the system fail to boot.