Added vmdk descriptor update
VMWare applications are made aware that an image has vm-tools installed
via tags appended to the VMDK's metadata. Typically, the tools write
the metadata during their first run, but in our case the first run of
the tools is the first boot of the appliance, so the host application
is not aware of tools until after a VM has been rebooted, missing a
critical opportunity to configure an instance before or during the
first boot. In order to allow VMWare applications (include vCloud)
to be aware that vm-tools are installed, we need to add the appropriate
metadata to the image during its creation