Deleted __checkMountDependencies runtime check
I'm afraid but this check causes more false positives than the
correct detection of an inconsistency in the mount table. Checking
if the string 'kiwi' appears in the process list is not enough,
especially in contained environment where the host kernel shares
mount information with a VM or container guest this check is
always wrong. We need a better solution here. Fixes Issue#501