Fixed JeOS templates configured for efi boot
in case of efi the efi modules should be installed to the system.
Even though it's not a problem for kiwi because it takes the modules
from the self build initrd it could become a problem when grub2
is used for reconfiguration on the system later. It is also
questionable whether it is a good idea from kiwi to install
the efi modules as part of the initrd just for the purpose to
take them from there at build time when they are not really needed
in the initrd at boot time. The new (next generation) kiwi version
will lookup modules from the system if possible and will not
keep them inside of the initrd