Added support for exfat as hybrid filesystem
In addition to fat also exfat is now supported as persistent
write filesystem. Because of the limitations of a fat filesystem
fat and exfat are only used as a container filesystem providing
an ext4 linux filesystem as a cowfile. The algorithm to create
the size of the cow file has also been changed to use half of
the size of the write partition or on fat a max size of 4G.
The size of the cowfile is also prepared to become overwritten.
However the XML definition and implementation to do this
is still missing