Fixed oem and vmx buildservice boot requires
The packages kiwi-desc-oemboot-requires and kiwi-desc-vmxboot-requires
did not explicitly require grub2 or the architecture specific efi module
packages for the arm architectures. The requirement was added by the
contents of the arm boot image descriptions. However those packages
will not be part of the boot image descriptions in the near future
and thus the buildservice needs them to be referenced explicitly
in the spec file