Delete vmxboot dracut optimization
For vmx type images, dracut was called in background to speedup
the boot process. However this could cause a race condition
together with grub2-mkconfig. If grub2-mkconfig is called but
dracut has not yet created the initrd, grub2 creates a configuration
file without an initrd. The result boot setup is not able to
reboot the system because the initrd is not loaded.
This fixes (bnc#982092)