Make sure netboot code populates kiwi_RootPart
For disk based images the file config.partids is written and
imported. However, for netboot stations a disk is optional
and no partids information exists. In order to provide at
least the partition number for the root partition on netboot
stations with a disk, the kiwi_RootPart is exported from
the provided client configuration file. Fixes bnc#1000194