UpdateOtherDeviceFstab method was not keeping the first argument
updateOtherDeviceFstab is method responsible to include the fstab
lines needed to mount additional partitions (apart from swap and
root) defined in the description file. This methid had an error
overwriting the variable $1 witht the value of $2, which is always
empty, as the method it is never used with more than a single
parameter. This fixes (bnc#1007765)