ActivateImage links correction and read-only test
This commit is related to bnc#1006834. Loading the iso image into
a read-only ramdisk caused some issues due to the activatImage
method cp commands. With this commit the symlinks are consistent
during the preinit phase and also before running the cp command
it is tested we are not in a read-only filesystem.