Testing isoinfo binary correctly
In method 'isols()' isoinfo command line tool is used, however the
binary was opened before testing if it existed. Apparently the open
perl function was not returning an error if the file did not exist.
So in the case that isoinfo was not located in /usr/bin/isoinfo the
open command was not failing, hence never trying to open the
fallback location /usr/lib/genisoimage/isoinfo.

This commit fixes #614.