Lazy umount for the chrooted environment
With this commit the mount points that fail to umount when trying a
regular umount, a lazy umount is applied. This can be helpful in
case the installation of some service in the chrooted environment
created/opened something inside a mount point preventing it to be
unmounted immediately.

This commit is related with the issue bnc#1008898