Fix vmdk disk convertion when using LVM
This patch fixes bsc#1059715. Calling 'pvs --noheadings -o vg_name <device>'
is not sufficient to get the volume name and to determine if LVM is being
used, as it could happen that udev events are not yet processed, thus the
<device> is not yet properly mapped. With this patch blkid is being used to
determine if the type of the requested partition is LVM and if so it proceeds
to a polling strategy to call pvs tool until it succeeds. Timeout is set to
30 seconds.