Do not support installation stick without boot partition
Installation stick bootloader configuration fails for the
installation image (installstick="true") if the image
is not configured to use a boot partition. Given the
fact the same functionallity can be achieved building
installation hybrid ISOs this specific configuration is not
supported.

This commit adds a runtime check to ensure this
combination is not set in the description file.

Fixes bsc#1107906