Normalize partedGetPartitionID output (bsc#1136744) (#675)
This commit ensures the partedGetPartitionID output returns the
same partition ID codes with or without the presence of sgdisk
utility. It basically trusts partition names provided by KIWI to map
partitions to specific IDs.

Additionally this commit also makes sure that the partition stated
to be of `ef` type in config.<mac> is named `legacy`, to be consistent
with the partedGetPartitionID expectations.