Improve partition validation for PXE
This commit includes a patch to relax the partition
size validation. It turned to happen that under some
specific partition size configurations in the PXE client
the partition layout was always seen as an invalid one
in each boot. This was causing a repartition and redeploy
of the PXE image in every single boot.

Fixes bsc#1141156