Disable filesystem unit tests
Unit tests regarding filesystem requires them to be installed
at build time according to the way the tests were written.
This is unfortunate and triggered a bunch of refactoring tasks
in the tests e.g on the fact that genisoimage does not exist
on sle15 as just one example. As this is all legacy code I
decided to go the easy way and just disable those tests