Do not match comments and PIs in XSLT templates
Adding toplevel comment blocks like it's needed for obs to
e.g enable the multipath feature made kiwi consume insane
amounts of memory during the XSLT transform step. While this
may be a bug in my version of lxml, we do not transform
comments on processing instructions in the conversion
templates, so the easiest solution is not to match them.